slight tweaks to interpreter

+++ tests/filecheck/interpreter/memref/storeandload.mlir
@@ -4,6 +4,6 @@ builtin.module {
     %0 = "arith.constant"() <{value = 42 : i32}> : () -> (i32)
     "memref.store"(%0, %m0) : (i32, memref<i32>) -> ()
     %1 = "memref.load"(%m0) : (memref<i32>) -> i32
-    func.return %1 : i64
+    func.return %1 : i32
   }
 }
